Make request file
