+++ frame/treekem/src/group_state.rs
@@ -7,8 +7,9 @@ use crate::ratchet_tree::{RatchetTree, RatchetTreeNode};
 use crate::serde::Serialize;
 use crate::store_path_secrets::StorePathSecrets;
 use crate::tree_math;
-use frame_common::crypto::{
-    ExportPathSecret, KeyVaultCmd, KeyVaultRequest, RecoverRequest, RecoveredPathSecret,
+use frame_common::{
+    crypto::{ExportPathSecret, RecoveredPathSecret},
+    request::{KeyVaultCmd, KeyVaultRequest, RecoverRequest},
 };
 use frame_mra_tls::{AttestedTlsConfig, Client, ClientConfig};
